% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{cleanup}
\alias{cleanup}
\title{Cleaning up PAT file}
\usage{
cleanup()
}
\value{
Boolean of cleanup success.
}
\description{
This function ensures that when having used a device, the access token is deleted.
This is important to ensure that there is no remainder lying around allowing illicit
access to your account.
}
