% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{clone_project}
\alias{clone_project}
\title{Clone OSF project to desktop}
\usage{
clone_project(id, path = getwd(), private = FALSE, maxdepth = 5)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX)}

\item{path}{a valid file path pointing to the location where the project will
be cloned. This is your current working directy by default.}

\item{private}{clone the project as seen privately?}

\item{maxdepth}{how many levels of subcomponents to trawl}
}
\value{
Boolean, clone success
}
\description{
This function copies an \emph{entire} project and its
components to the harddrive of the individual
(depth of clone depends on the maxdepth argument).
There is currently no way to estimate the size
before downloading so it might take a while. BUT
there's a progess bar :-) Currently limited to only
files stored on OSF (not via add-ons).
}
\examples{
\dontrun{
clone_project(id = "12345")}
}
