% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{comment.osf}
\alias{comment.osf}
\title{Comment on a project on the OSF}
\usage{
comment.osf(id = NULL, txt = NULL, ...)
}
\arguments{
\item{id}{OSF id as depicted in link by XXXX, osf.io/XXXX}

\item{txt}{Comment text}
}
\description{
Comment on a project on the OSF
}

