% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{comment}
\alias{comment}
\title{Comment on a project on the OSF}
\usage{
comment(id = NULL, txt = NULL, ...)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX)}

\item{txt}{Contents of the comment}
}
\value{
Boolean, posting succeeded?
}
\description{
Comment on a project on the OSF
}
\examples{
\dontrun{comment(id = '12345', txt = 'This is an example')}
}

