% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/construct_link.R
\name{construct.link}
\alias{construct.link}
\title{Construct an API link with proper base}
\usage{
construct.link(request = NULL)
}
\arguments{
\item{request}{The request link to be combined with the base API link.}
}
\value{
The full request link with proper base
}
\description{
Construct an API link with proper base
}
\examples{
construct.link("nodes/NODE_ID/files/")
}

