% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_link.R
\name{construct_link_files}
\alias{construct_link_files}
\title{Construct a waterbutler API link with proper base}
\usage{
construct_link_files(id = NULL, request = NULL, test = FALSE)
}
\arguments{
\item{id}{OSF id}

\item{request}{Request for waterbutler}

\item{test}{Whether the link should be constructed for testing server}
}
\value{
Waterbutler link
}
\description{
Construct a waterbutler API link with proper base
}

