% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{create_component}
\alias{create_component}
\title{Create a component within a project}
\usage{
create_component(id, title, description = "", category = "project",
  private = TRUE)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX) of parent project}

\item{title}{Title of the component [required]}

\item{description}{Description of the component [optional]}

\item{category}{Category of component, for valid categories
see \code{\link{process_category}}, defaults to `project`}

\item{private}{Boolean of whether the component is supposed to be private}
}
\value{
OSF id of created component
}
\description{
This function creates a new component within a project. New components can
be created as private or public components using the \code{private} argument.
}
\examples{
\dontrun{
create_component("12345", "New Test Component")
}
}
\seealso{
\code{\link{create_project}}
}
