% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{delete.nodes}
\alias{delete.nodes}
\title{Delete a node with its id}
\usage{
delete.nodes(id)
}
\arguments{
\item{id}{The node_id to be deleted.}
}
\value{
1
}
\description{
Delete a node with its id
}

