% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Delete a project or file from the OSF}
\usage{
delete(id = NULL, recursive = FALSE, maxdepth = 5, ...)
}
\arguments{
\item{id}{OSF id (osf.io/xxxx)}

\item{recursive}{Boolean, if TRUE will go through folder nesting (see \code{maxdepth})}

\item{maxdepth}{Number of nesting levels to go through}

\item{\ldots}{Additional parameters passed to \code{\link{process_type}}, \code{\link{delete_project}}, and \code{\link{delete_file}}}
}
\value{
Boolean, deletion succeeded?
}
\description{
Delete a project or file from the OSF
}
