% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{delete_component}
\alias{delete_component}
\title{Empty out a component and delete it}
\usage{
delete_component(id)
}
\arguments{
\item{id}{OSF id (osf.io/xxxx; just XXXX)}
}
\value{
Boolean, deletion success
}
\description{
Empty out a component and delete it
}
