% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_files}
\alias{delete_files}
\title{Delete a file based on OSF id}
\usage{
delete_files(id)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX)}
}
\value{
Boolean, delete succeeded?
}
\description{
Delete a file based on OSF id
}
