% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{delete_folder}
\alias{delete_folder}
\title{Delete a folder on the OSF}
\usage{
delete_folder(url)
}
\arguments{
\item{url}{Waterbutler link}
}
\value{
Boolean, deletion success?
}
\description{
Delete a folder on the OSF
}
\seealso{
\code{\link{create_folder}}
}
