% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes-delete.R
\name{delete_project}
\alias{delete_project}
\alias{delete_component}
\title{Delete a project or component}
\usage{
delete_project(id, recursive = FALSE)

delete_component(id, recursive = FALSE)
}
\arguments{
\item{id}{the OSF entity's unique identifier}

\item{recursive}{remove sub-components before deleting the top-level entity}
}
\value{
the deleted entity's unique identifier (invisibly)
}
\description{
Functions to \strong{permanently} delete a project or component from OSF,
including any uploaded files, wiki content, or comments contained therein.
This process does not request confirmation, so please \strong{handle with
care}. This functionality is limited to contributors with admin-level
permissions.
}
\details{
If the project or component contains sub-components, those must be deleted
first. Setting \code{recursive = TRUE} will attempt to remove the hierarchy
of sub-components before deleting the top-level entity.
}
\examples{
\dontrun{
delete_project(id = "y7w8p")}
}
