% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.osf}
\alias{download.osf}
\title{Download files from the OSF}
\usage{
download.osf(id = NULL, file = NULL, private = FALSE)
}
\arguments{
\item{id}{Specify the node id (osf.io/XXXX)}

\item{file}{Specify path to save file. If NULL, defaults to OSF}

\item{private}{Boolean to specify whether file is private}
}
\description{
Download files from the OSF
}
\examples{
download.osf('zevw2', 'test123.md')
}

