% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{download_file}
\alias{download_file}
\title{Download files from the OSF}
\usage{
download_file(id, path = NULL, private = FALSE)
}
\arguments{
\item{id}{Specify the node id (osf.io/XXXX)}

\item{path}{Specify path to save file to. If NULL, defaults to OSF filename in \code{\link{tempdir}}}

\item{private}{Boolean to specify whether file is private}
}
\value{
Return filepath for easy processing
}
\description{
Download files from the OSF
}
\examples{
\dontrun{
download_file('zevw2', 'test123.md')
}
}
