% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{get.nodes}
\alias{get.nodes}
\title{Retrieve nodes associated with id}
\usage{
get.nodes(id = NULL, contributors = FALSE, files = FALSE,
  children = FALSE)
}
\arguments{
\item{id}{The id to search for. Use `NULL` to retrieve all,
`me` for logged in account. Maximum of 1 id.}

\item{contributors}{Boolean to extract the contributors of the node}

\item{files}{Boolean to download files}

\item{children}{Boolean to retrieve children nodes of id}
}
\value{
List object of results
}
\description{
Retrieve nodes associated with id
}
\examples{
get.nodes()
get.nodes(id = 'me')
get.nodes(id = 'nu97z')
}

