% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_users.R
\name{get.users}
\alias{get.users}
\title{Return information on an OSF user}
\usage{
get.users(id = NULL, nodes = FALSE)
}
\arguments{
\item{id}{Argument to specify the OSF id ('me' refers to logged in account)}

\item{nodes}{Boolean to return the nodes available for that user}
}
\value{
List object with account information
}
\description{
Return information on an OSF user
}

