% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config}
\alias{get_config}
\title{Create authorization config}
\usage{
get_config(login)
}
\arguments{
\item{login}{Boolean indicating whether login is required.}
}
\value{
httr::request configuration
}
\description{
Create authorization config
}
