% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file_info}
\alias{get_file_info}
\title{Get data frame of information about all files in an OSF node}
\usage{
get_file_info(id, private = FALSE)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) for the node (project or component) to get file info for}

\item{private}{Boolean to specify whether to get info for private files}
}
\description{
Get data frame of information about all files in an OSF node
}
