% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Retrieve nodes associated with id}
\usage{
get_nodes(id = NULL, contributors = FALSE, files = FALSE,
  children = FALSE, private = FALSE)
}
\arguments{
\item{id}{The id to search for. Use `NULL` to retrieve all,
`me` for logged in account. Maximum of 1 id.}

\item{contributors}{Boolean to extract the contributors of the node}

\item{files}{Boolean to retrieve files}

\item{children}{Boolean to retrieve children nodes of id}

\item{private}{Boolean, retrieve private node}
}
\value{
List object of results
}
\description{
Retrieve nodes associated with id
}
\examples{
\dontrun{get_nodes()
get_nodes(id = 'me')
get_nodes(id = 'nu97z')
}
}
