% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users}
\alias{get_users}
\title{Return information on an OSF user}
\usage{
get_users(id = "me", nodes = FALSE)
}
\arguments{
\item{id}{String, OSF id (defaults to 'me'; logged in account)}

\item{nodes}{Boolean, return nodes available for that user?}
}
\value{
List object with account information
}
\description{
Return information on an OSF user
}
