% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{http_error}
\alias{http_error}
\title{Stop execution with HTTP status code}
\usage{
http_error(code, ...)
}
\arguments{
\item{code}{HTTP status code}

\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}
}
\description{
Stop execution with HTTP status code
}
