% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logged_in}
\alias{logged_in}
\title{Identify whether user is logged in}
\usage{
logged_in()
}
\value{
Boolean indicating logged in status
}
\description{
This function identifies whether the user has logged in by providing a
personal access token (PAT).
}
