% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{login}
\alias{login}
\title{Login function; interactive without arguments}
\usage{
login(pat = NULL)
}
\arguments{
\item{pat}{Personal Access Token (PAT) for fast login. If no pat is given, function queries for it.}
}
\value{
Personal access token from global environment.
}
\description{
Function to easily login, either with a prompt when run without arguments, or
without prompt when run with arguments or config file (~/.osf_config). If you
run this function without arguments, the prompt will result in creation of a
~/.osf_config file (in your Documents folder for Windows).
}
\details{
For the security of your OSF account, do not forget to delete this file. This
can be done with the logout() function as well.
}
