% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{logout}
\alias{logout}
\title{Logout function}
\usage{
logout()
}
\value{
Boolean succes of logout
}
\description{
Log out of the OSF and clean up after yourself for your own security. This 
ensures that you do not leave a personal access token lying around on the 
computer for others to abuse without your knowledge (before revoking it, 
but then the damage is already done).
}
