% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_file.R
\name{move_file}
\alias{move_file}
\title{Move (and copy) files on the OSF}
\usage{
move_file(from = NULL, to = NULL, filename = NULL, action = "move",
  conflict = "replace", ...)
}
\arguments{
\item{from}{OSF file id to move (osf.io/xxxx)}

\item{to}{OSF id to move to (osf.io/xxxx; needs to be component)}

\item{filename}{Optional, rename the file}

\item{action}{Move or copy}

\item{conflict}{Keep old file or replace in case of conflict}
}
\value{
Boolean, moving succeeded?
}
\description{
Move (and copy) files on the OSF
}

