% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{move_files}
\alias{move_files}
\title{Move (and copy) files on OSF}
\usage{
move_files(from = NULL, to = NULL, filename = NULL, action = "move",
  conflict = "replace")
}
\arguments{
\item{from}{OSF file id to move (osf.io/XXXXX; just XXXXX)}

\item{to}{OSF id to move to (osf.io/XXXXX; needs to be component)}

\item{filename}{Optional, rename the file}

\item{action}{'move' or 'copy'}

\item{conflict}{Keep old file or replace in case of conflict}
}
\value{
Boolean, moving succeeded? (invisible)
}
\description{
Move (and copy) files on OSF
}
