% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-create.R
\name{node-create}
\alias{node-create}
\alias{osf_project_create}
\alias{osf_component_create}
\title{Create a new OSF project or component}
\usage{
osf_project_create(title = NULL, description = "", private = TRUE)

osf_component_create(id, title = NULL, description = "",
  private = TRUE)
}
\arguments{
\item{title, description}{Set a title (required) and, optionally, a description}

\item{private}{Boolean, should it be private (defaults to \code{TRUE})}

\item{id}{GUID of the existing OSF project or component in which the new
component will be created}
}
\value{
An object of class \code{osf_tbl_node} with one row per item.
}
\description{
Create a new OSF project or component
}
\examples{
\dontrun{
# create a new project
proj <- osf_project(title = "New Private OSF Project")

# update/retrieve project details
osf_project(proj$id)
}
}
