% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-update.R
\name{node-update}
\alias{node-update}
\alias{osf_project_update}
\alias{osf_component_update}
\title{Update an OSF project or component}
\usage{
osf_project_update(id, title = NULL, description = NULL,
  private = NULL)

osf_component_update(id, title = NULL, description = NULL,
  private = NULL)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX)}

\item{private}{Set project to private/public (default changes to public)}
}
\value{
Boolean, update success
}
\description{
This function updates the private/public status of a component. By default,
the function turns a private component into a public component. This
function simply wraps the \code{\link{update_project}} function
as it has the same operations.
}
\examples{
\dontrun{
proj <- osf_project_create("temporary-title", private = TRUE)
proj <- osf_project_update(proj$id, title = "Final Title", private = FALSE)
}
}
\seealso{
\code{\link{update_project}}
}
