% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_auth.R
\name{osf_auth}
\alias{osf_auth}
\title{Authenticate osfr with a personal access token}
\usage{
osf_auth(token = NULL)
}
\arguments{
\item{token}{OSF personal access token.}
}
\value{
Invisibly returns your OSF PAT along with a message indicating it was
registered.
}
\description{
Authorize osfr to interact with your OSF data and OSF account by passing a
personal access token (PAT) to \code{osf_auth()}. If no \code{token} is provided,
\code{osf_auth()} will attempt to obtain a PAT from the \code{OSF_PAT} environment
variable. However, since osfr checks for the presence of \code{OSF_PAT} on
start-up, this is only necessary if the variable was created or redefined in
the middle of a session. See below for additional details and instructions
for generating and utilizing your PAT.
}
\details{
Out of the box osfr can only access publicly available projects, components,
and files on OSF. In order for osfr to view and manage your private resources
you must provide a Personal Access Token (PAT). The following instructions
will walk you through the process of generating a PAT and using it to
authenticate osfr.
}
\section{Creating an OSF personal access token}{

\itemize{
\item Navigate to \url{https://osf.io/settings/tokens/}
\item Click the \emph{New token} button and provide a descriptive name
\item Select the scopes (i.e., permissions) you'd like to grant osfr
\item Click the \emph{Create} button to generate your PAT
\item If successful, your 70 character token will be displayed along with several
important warnings you should definitely read over carefully
\item You read those warnings, right?
\item Copy your token and keep it in a safe place
}
}

\section{Using your PAT with osfr}{


There are two possible approaches for authenticating osfr with your PAT.

The simpler approach is to call the \code{osf_auth()} function at the start of
every new R session and manually paste in your token. Note that your PAT
should be treated like a password and, as such, should not be hardcoded into
your script.

A better approach is to store your PAT as an environment variable called
\code{OSF_PAT}. Doing so will allow osfr to detect and utilize the token
automatically without any need to manually call \code{osf_auth()}. One way to
accomplish this is by creating an \code{.Renviron} file in your home or working
directory that defines the \code{OSF_PAT} variable. For example:\preformatted{OSF_PAT=bdEEFMCuBtaBoSK11YzyjOdjUjKtWIj2FWxHl6kTBRax7uaeyBghALumTO1kT8RA
}

For new users we suggest adding the \code{.Renviron} to your home directory so it
is automatically applied to all your projects. To verify this was done
correctly, restart R and run \code{Sys.getenv("OSF_PAT")}, which should return
your PAT.
}

\examples{
\dontrun{
# manually authenticate with a PAT
osf_auth("bdEEFMCuBtaBoSK11YzyjOdjUjKtWIj2FWxHl6kTBRax7uaeyBghALumTO1kT8RA")
}

}
\references{
\enumerate{
\item Colin Gillespie and Robin Lovelace (2017). \emph{Efficient R programming}.
O'Reilly Press. \url{https://csgillespie.github.io/efficientR}.
}
}
