% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_create.R
\name{osf_create}
\alias{osf_create}
\alias{osf_create_project}
\alias{osf_create_component}
\title{Create a new project or component on OSF}
\usage{
osf_create_project(title, description = NULL, public = FALSE)

osf_create_component(x, title, description = NULL, public = FALSE)
}
\arguments{
\item{title, description}{Set a title (required) and, optionally, a
description.}

\item{public}{Logical, should it be publicly available (\code{TRUE}) or private
(\code{FALSE}, the default)?}

\item{x}{An \code{\link{osf_tbl_node}} with a single OSF project or component that will
serve as the new sub-component's parent node.}
}
\value{
an \code{\link{osf_tbl_node}} containing the new OSF project or component
}
\description{
Use \code{osf_create_project()} to create a new top-level project on OSF. A nested
component can be created by providing an \code{\link{osf_tbl_node}} referencing an
existing project or component to \code{osf_create_component()}'s \code{x} argument.
}
\section{}{
 OSF Nodes
Both projects and components are implemented as \emph{nodes} within OSF, which
explains why they are functionally identical. The only distinction between
the two is projects exist at the top-level of an organizational hierarchy,
and components are nested within a parent project or another component. As
such, osfr uses the same S3 class to represent both projects and components:
an \code{\link{osf_tbl_node}}.
}

\examples{
\dontrun{
# create a new private project
project <- osf_create_project(title = "Private OSF Project")

# create a new component
component <- osf_create_component(project, title = "Project Data")
}
}
\references{
\enumerate{
\item OSF Guides: Create a Project.
\url{http://help.osf.io/m/projects/l/481539-create-a-project}.
\item OSF Guides: Create a Component.
\url{http://help.osf.io/m/projects/l/481998-create-components}
}
}
