% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_download.R
\name{osf_download}
\alias{osf_download}
\title{Download an OSF file or folder}
\usage{
osf_download(x, path = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{an \code{\link{osf_tbl_file}} containing a file or directory}

\item{path}{Local path where the downloaded file will be saved. The default
is to use the remote file's name.}

\item{overwrite}{Logical, if the local path already exists should it be
replaced with the downloaded file?}
}
\value{
\code{TRUE}, invisibly if download succeeded
}
\description{
Files stored in OSF projects, components, or folders can be downloaded
locally by providing the appropriate \code{\link{osf_tbl_file}} to \code{osf_download()}.
Note that directories are downloaded as zip files.
}
\examples{
\dontrun{
analysis_plan <- osf_retrieve_file("79b2n")
osf_download(analysis_plan)
}
}
