% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{osf_ls}
\alias{osf_ls}
\title{List files and folders}
\usage{
osf_ls(id, path = NULL, path_id = NULL, type = "any",
  pattern = NULL, n_max = Inf)
}
\arguments{
\item{id}{OSF project/component GUID}

\item{path}{list files within the specified subdirectory path}

\item{path_id}{OSF unique identifier assigned to a directory}

\item{type}{filter response to include only \code{"files"} or
\code{"folders"}}

\item{pattern}{filter response to include entities whose name contain the
specified pattern}
}
\description{
List the files and folders in the top-level of an OSF Project or Component.
}
\details{
Specify a \code{path} or a \code{path_id} to list the contents of a
particular subdirectory. If both \code{path} \emph{and} \code{path_id} are
defined, then \code{path} is assumed to be a subdirectory within the
directory corresponding to the supplied \code{path_id}.
}
