% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_ls_files.R
\name{osf_ls_files}
\alias{osf_ls_files}
\title{List files and folders}
\usage{
osf_ls_files(x, path = NULL, type = "any", pattern = NULL,
  n_max = 10)
}
\arguments{
\item{x}{an \code{\link{osf_tbl_node}} representing an OSF project or component or an
\code{\link{osf_tbl_file}} containing a directory}

\item{path}{list files within the specified subdirectory path}

\item{type}{filter query to include only \code{"file"} or \code{"folder"}}

\item{pattern}{filter query for entities whose name contains the
specified \code{pattern}}

\item{n_max}{Number of results to return from OSF.}
}
\value{
an \code{\link{osf_tbl_file}} with \code{x}'s top-level files and directories
}
\description{
List the files and folders in the top-level of an OSF project, component, or
directory. Specify a \code{path} to list the contents of a particular
subdirectory.
}
