% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_ls_preprints.R
\name{osf_ls_preprints}
\alias{osf_ls_preprints}
\title{List preprints from a specific user}
\usage{
osf_ls_preprints(x, pattern = NULL, n_max = 10, verbose = FALSE)
}
\arguments{
\item{x}{one of the following:
\itemize{
\item An \code{\link{osf_tbl_user}} with a single OSF user.
}}

\item{pattern}{Character string used to filter for results that contain the
substring \code{"pattern"} in their name. \emph{Note:} this is a fixed, case-insensitive
search.}

\item{n_max}{Maximum number of results to return from OSF (default is 10).
Set to \code{Inf} to return \emph{all} results.}

\item{verbose}{Logical, indicating whether to print informative messages
about interactions with the OSF API (default \code{FALSE}).}
}
\value{
An \code{\link{osf_tbl_node}} with one row for each OSF preprint
ordered by modification time.
}
\description{
List the preprints that are associated with a specific user.
Note that will return \emph{all} preprints, unlike \code{\link[=osf_ls_nodes]{osf_ls_nodes()}}.
}
\examples{
\dontrun{
# List your recent projects and components
osf_retrieve_user("me") \%>\%
  osf_ls_preprints()
}
}
\seealso{
\code{\link[=osf_ls_nodes]{osf_ls_nodes()}} to generate a list of all project components.
}
