% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_refresh.R
\name{osf_refresh}
\alias{osf_refresh}
\title{Refresh an OSF entity}
\usage{
osf_refresh(x)
}
\arguments{
\item{x}{an \code{\link{osf_tbl}}.}
}
\description{
Use \code{osf_refresh()} to update one or more entities in an \code{\link[=osf_tbl]{osf_tbl()}} with
the latest information from OSF.
}
