% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{osf_user_retrieve}
\alias{osf_user_retrieve}
\title{Retrieve OSF user information}
\usage{
osf_user_retrieve(id = "me")
}
\arguments{
\item{id}{A valid OSF GUID}
}
\description{
Retrieve OSF user information
}
