% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osfr-package.R
\docType{package}
\name{osfr-package}
\alias{osfr}
\alias{osfr-package}
\title{osfr: R interface to OSF}
\description{
osfr provides a suite of functions for interacting with
\href{https://www.osf.io}{OSF} that are primarily focused on project management
workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://centerforopenscience.github.io/osfr}
  \item \url{https://github.com/centerforopenscience/osfr}
  \item Report bugs at \url{https://github.com/centerforopenscience/osfr/issues}
}

}
\author{
\strong{Maintainer}: Aaron Wolen \email{aaron@wolen.com} (0000-0003-2542-2202)

Authors:
\itemize{
  \item Chris Hartgerink \email{chjh@protonmail.com} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID})
}

Other contributors:
\itemize{
  \item Timothy York \email{timothypyork@gmail.com} (\href{https://orcid.org/0000-0003-4068-4286}{ORCID}) [contributor]
  \item Ryan Hafen \email{rhafen@purdue.edu} [contributor]
  \item Brian Richards \email{brian.g.richards@gmail.com} [contributor]
  \item Courtney Soderberg \email{courtney@cos.io} (\href{https://orcid.org/0000-0003-1227-7042}{ORCID}) [contributor]
}

}
