% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osfr.R
\docType{package}
\name{osfr}
\alias{osfr}
\alias{osfr-package}
\title{osfr: A package for interacting with OSF from the Center for Open Science.}
\description{
The osfr package helps you document your research as you go by creating
projects, uploading and downloading files, and much more.
}
\details{
Suggestions for improvements or bug reports are always welcome at
\url{https://github.com/CenterForOpenScience/osfr}
}
