% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_users.R
\name{patch.users}
\alias{patch.users}
\title{Patch user information}
\usage{
patch.users(id = "me", full_name = NULL, given_name = NULL,
  middle_names = NULL, family_name = NULL, suffix = NULL)
}
\arguments{
\item{suffix}{}
}
\value{
Boolean of success of patching user information
}
\description{
Patch user information
}

