% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{post.nodes}
\alias{post.nodes}
\title{Create a new project}
\usage{
post.nodes(type = "nodes", title = NULL, description = NULL,
  category = "project", tags = NULL, public = TRUE)
}
\arguments{
\item{type}{Redundant argument, required by API only.}

\item{title}{String of project title}

\item{description}{String, description of project (optional)}

\item{category}{String, type of node, options under details}

\item{tags}{Vector of strings, tags to be posted on project}

\item{public}{Boolean, sets project to public or private; defaults to public}

\item{user}{String, login username}

\item{password}{String, login password}
}
\description{
Create a new project
}

