% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{private.nodes}
\alias{private.nodes}
\title{Make node and all subnodes private}
\usage{
private.nodes(id = NULL)
}
\arguments{
\item{id}{node to crawl and make private, including subnodes}
}
\description{
Make node and all subnodes private
}

