% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_category}
\alias{process_category}
\title{Processing whether a category is valid}
\usage{
process_category(category = "")
}
\arguments{
\item{category}{Category to check for validity. Valid categories:
\itemize{
  \item project
  \item hypothesis
  \item methods and measures
  \item procedure
  \item instrumentation
  \item data
  \item analysis
  \item communication
  \item other
}}
}
\description{
Processing whether a category is valid
}
