% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-waterbutler-api.R
\name{process_file_id}
\alias{process_file_id}
\title{Processing a file id to waterbutler}
\usage{
process_file_id(id, private = FALSE)
}
\arguments{
\item{id}{OSF id (osf.io/xxxx; just xxxx)}

\item{private}{Boolean, if file is private}
}
\value{
Waterbutler URL
}
\description{
The OSF uses a separate API to handle file requests, for some reason. This
function helps deal with that.
}
