% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_json}
\alias{process_json}
\title{Function to parse API call}
\usage{
process_json(x)
}
\arguments{
\item{x}{Object containing the result of an API call.}
}
\value{
Parsed JSON object in the form of an R object.
}
\description{
Function to parse API call
}
