% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_pagination}
\alias{process_pagination}
\title{Process Pagination}
\usage{
process_pagination(res, config)
}
\arguments{
\item{res}{The initial list return from the OSF API and run through
`osfr::process_json()`. Must contain the links section.}

\item{config}{The configuration used in the initial call to the OSF API.}
}
\value{
List of all of the pages from the API (including the input list).
}
\description{
Processes the paginated data returned by the OSF and returns a list with
all of the pages combined.
}
