% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_response.R
\name{process_response}
\alias{process_response}
\title{Process a response from the OSF API}
\usage{
process_response(res)
}
\arguments{
\item{res}{}
}
\value{
list
}
\description{
Process a response from the OSF API
}
