% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_type}
\alias{process_type}
\title{Identify type of endpoint for id}
\usage{
process_type(id)
}
\arguments{
\item{id}{OSF id to check (osf.io/xxxxx).}
}
\value{
Endpoint of id as character (nodes | files)
}
\description{
This function takes an OSF id and returns the type of endpoint for that id.
Only returns types for the nodes and files endpoints.
}
