% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-waterbutler-api.R
\name{process_waterbutler}
\alias{process_waterbutler}
\title{Materialize Waterbutler URL}
\usage{
process_waterbutler(url, private = TRUE)
}
\arguments{
\item{url}{Waterbutler URL, starts with files.osf.io}

\item{private}{Boolean, whether or not the file is prvate}
}
\value{
OSF id
}
\description{
Materialize Waterbutler URL
}
