% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{public.nodes}
\alias{public.nodes}
\title{Make node and all subnodes public}
\usage{
public.nodes(id = NULL)
}
\arguments{
\item{id}{node to crawl and make public, including subnodes}
}
\description{
Make node and all subnodes public
}

