% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_users.R
\name{put.users}
\alias{put.users}
\title{Put user information}
\usage{
put.users(id = "me", type = "users", full_name = NULL,
  given_name = NULL, middle_names = NULL, family_name = NULL,
  suffix = NULL)
}
\arguments{
\item{id}{the id to put information for}

\item{type}{type of id, 'users' only option}

\item{suffix}{}
}
\value{
Boolean indicating success of PUTting user information
}
\description{
Put user information
}

