% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{recurse.nodes}
\alias{recurse.nodes}
\title{Function to crawl node for all sub-ids.}
\usage{
recurse.nodes(id = NULL, login = FALSE)
}
\arguments{
\item{id}{Parent node to crawl.}

\item{login}{Boolean, indicates whether crawl should occur with login or without.}
}
\value{
Vector of node ids, first one is always the parent entered as id
}
\description{
Function to crawl node for all sub-ids.
}

