% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse_node.R
\name{recurse_node}
\alias{recurse_node}
\title{Function to crawl through OSF project}
\usage{
recurse_node(id = NULL, public = TRUE, maxdepth = 5, ...)
}
\arguments{
\item{id}{OSF parent ID (osf.io/xxxx) to crawl}

\item{public}{Boolean, search for only public?}

\item{maxdepth}{Integer, amount of levels deep to crawl}

\item{\ldots}{Additional parameters passed to \code{\link{construct_link}}}
}
\value{
List of OSF ids, with parents as very last.
}
\description{
Function to crawl through OSF project
}
