% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_space}
\alias{rm_space}
\title{Function to correct link into something parsable}
\usage{
rm_space(x)
}
\arguments{
\item{x}{String to correct}
}
\value{
Parsable link
}
\description{
Function to correct link into something parsable
}
