% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.osf}
\alias{search.osf}
\title{Searching the OSF}
\usage{
search.osf(type = "nodes", ... = NULL)
}
\arguments{
\item{type}{Specifying what type of information to search}

\item{...}{Any arguments from the search.nodes or search.users functions}
}
\description{
Searching the OSF
}
\examples{
search.osf(title = 'many labs', type = 'nodes')
}

