% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_osf}
\alias{search_osf}
\title{Searching the OSF}
\usage{
search_osf(type = "nodes", ...)
}
\arguments{
\item{type}{Specifying what type of information to search}

\item{...}{Any arguments from the search.nodes or search.users functions}
}
\value{
Data frame of nodes or users
}
\description{
Searching the OSF
}
\examples{
\dontrun{search.osf(title = 'many labs', type = 'nodes')}
}
