% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_users}
\alias{search_users}
\title{Search OSF users}
\usage{
search_users(full_name = NULL, family_name = NULL, ...)
}
\arguments{
\item{full_name}{Search the entire name}

\item{family_name}{Search just the family name (full_name encompasses more)}
}
\value{
Data frame of users
}
\description{
Search OSF users
}

