% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{update_component}
\alias{update_component}
\title{Update a component on the OSF}
\usage{
update_component(id, private = FALSE)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX)}

\item{private}{Set project to private/public (default changes to public)}
}
\value{
Boolean, update success
}
\description{
Simply wraps the \code{\link{update_project}} function
because it has the same operations.
}
\seealso{
\code{\link{update_project}}
}
