% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{update_project}
\alias{update_project}
\title{Update an OSF project}
\usage{
update_project(id, private = FALSE)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX)}

\item{private}{Set project to private/public (default changes to public)}
}
\value{
Boolean of update success
}
\description{
This updates some of the metadata for an OSF project.
Currently limited to switching a project to public.
Does not yet include making public all subcomponents
or vectorized to make multiple projects public at the
same time.
}
\examples{
\dontrun{
update_project(id = "12345")}
}
