% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_revised_files}
\alias{upload_revised_files}
\title{Upload a revised file to OSF}
\usage{
upload_revised_files(id, path)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX) of file to revise}

\item{path}{Path to file on local machine to upload.}
}
\value{
Boolean, revision success? (invisible)
}
\description{
Upload a revised file to OSF
}
\seealso{
\code{\link{upload_files}}, \code{\link{upload_new_files}}
}
