% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_revision}
\alias{upload_revision}
\title{Upload a revised file to the OSF}
\usage{
upload_revision(id, path)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX) of file to revise}

\item{path}{Path to file on local machine to upload.}
}
\value{
Boolean, revision success? (invisible)
}
\description{
Upload a revised file to the OSF
}
\seealso{
\code{\link{upload_files}}, \code{\link{upload_new}}
}
