% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact_users.R
\name{users.all}
\alias{users.all}
\title{Retrieve all users and their information from the OSF API}
\usage{
users.all()
}
\value{
Object dataframe including, for each user:
\enumerate{
\item id
\item fullname
\item given_name
\item middle_names
\item family_name
\item suffix
\item date_registered
\item profile_image_url
\item gitHub
\item scholar
\item personal_website
\item twitter
\item linkedIn
\item impactStory
\item orcid
\item researcherId
\item links
\item nodes
}
}
\description{
Retrieve all users and their information from the OSF API
}

