% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{view_project}
\alias{view_project}
\title{View an OSF project on osf.io}
\usage{
view_project(id)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX)}
}
\description{
This function opens the project in a web browser. If the project is private,
you will be asked to log into OSF from the browser.
}
\examples{
\dontrun{
view_project(id = "m5pds")}
}
