% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{welcome}
\alias{welcome}
\title{Personalised welcome}
\usage{
welcome()
}
\value{
Welcome message of logged in user, if any
}
\description{
Function that helps you identify who's logged in, just in case you forget.
}
