% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{comment_osf}
\alias{comment_osf}
\title{Comment on a project on the OSF}
\usage{
comment_osf(id = NULL, txt)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX)}

\item{txt}{Contents of the comment}
}
\value{
Boolean, posting succeeded?
}
\description{
Provide comments on an OSF project. Please be respectful and inclusive in doing
so and see the Terms of Use of the OSF.
\url{https://github.com/CenterForOpenScience/cos.io/blob/master/TERMS_OF_USE.md}
}
\examples{
\dontrun{comment_osf(id = '12345', txt = 'This is an example')}
}
