% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{comment_osf}
\alias{comment_osf}
\title{Comment on a project on the OSF}
\usage{
comment_osf(id = NULL, txt = NULL, ...)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX)}

\item{txt}{Contents of the comment}

\item{\ldots}{Additional parameters passed to \code{\link{process_type}} and \code{\link{construct_link}}}
}
\value{
Boolean, posting succeeded?
}
\description{
Comment on a project on the OSF
}
\examples{
\dontrun{comment_osf(id = '12345', txt = 'This is an example')}
}
