% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{construct.link}
\alias{construct.link}
\title{Construct an API link with proper base}
\usage{
construct.link(request = NULL, login = FALSE)
}
\arguments{
\item{request}{The request link to be combined with the base API link.}
}
\value{
The full request link with proper base
}
\description{
Construct an API link with proper base
}
\examples{
construct.link("nodes/{node_id}/files/")
}

