% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-osf-api.R
\name{construct_link}
\alias{construct_link}
\title{Construct an API link with proper base}
\usage{
construct_link(request)
}
\arguments{
\item{request}{The request link to be combined with the base API link.}
}
\value{
The full request link with proper base.
}
\description{
Building urls for requests. OSF_USE_SERVER can be specified in the
environment to use a test or staging server.
}
