% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_link_files}
\alias{construct_link_files}
\title{Construct a waterbutler API link with proper base}
\usage{
construct_link_files(id = NULL, provider = "osfstorage", request = NULL)
}
\arguments{
\item{id}{OSF id}

\item{provider}{Storage provider (osfstorage, github, etc)}

\item{request}{Request for waterbutler}
}
\value{
Waterbutler link
}
\description{
Construct a waterbutler API link with proper base
}
