% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{create_component}
\alias{create_component}
\title{Create a component within a project}
\usage{
create_component(id, title = "", description = "", category = "",
  private = TRUE)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX) of parent project}

\item{title}{Title of the component}

\item{description}{Description of the component}

\item{category}{Category of component, for valid categories
see \code{\link{process_category}}}

\item{private}{Boolean of whether the component is supposed to be private}
}
\value{
OSF id of created component
}
\description{
Create a component within a project
}
\seealso{
\code{\link{create_project}}
}
