% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_component.R
\name{create_component}
\alias{create_component}
\title{Create a component within a project}
\usage{
create_component(id = NULL, title = "", description = "", category = "",
  public = TRUE, ...)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) of parent project}

\item{title}{Title of the component}

\item{description}{Description of the component}

\item{category}{Category of component, for valid categories
see \code{\link{process_category}}}

\item{public}{Boolean of whether the component is supposed to be public}
}
\value{
OSF id of created component
}
\description{
Create a component within a project
}
\seealso{
\code{\link{create_project}}
}

