% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{create_folder}
\alias{create_folder}
\title{Upload a new file to the OSF.}
\usage{
create_folder(id, path)
}
\arguments{
\item{id}{Parent OSF project id (osf.io/XXXX) to upload to.}

\item{path}{Path of the folder to create on OSF.}
}
\value{
Waterbutler URL
}
\description{
Upload a new file to the OSF.
}
