% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a project on the OSF}
\usage{
create_project(title = "", description = "", public = TRUE, ...)
}
\arguments{
\item{title}{Project title}

\item{description}{Project description}

\item{public}{Boolean, whether project should be public (FALSE for private)}

\item{\ldots}{Additional parameters passed to \code{\link{construct_link}}}
}
\value{
Returns the created project's id
}
\description{
Create a project on the OSF
}
\examples{
\dontrun{
create_project(title = "Testing the OSF project creation")
}
}
\seealso{
\code{\link{create_component}}
}
