% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{create_project}
\alias{create_project}
\title{Create a project on the OSF}
\usage{
create_project(title = "", description = "", private = TRUE)
}
\arguments{
\item{title}{Project title}

\item{description}{Project description}

\item{private}{Boolean, whether project should be private}
}
\value{
Returns the created project's id
}
\description{
Create a project on the OSF
}
\examples{
\dontrun{
create_project(title = "Testing the OSF project creation")
}
}
\seealso{
\code{\link{create_component}}
}
