% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{delete.nodes}
\alias{delete.nodes}
\title{Delete a node with its id}
\usage{
delete.nodes(id = NULL, recursive = FALSE)
}
\arguments{
\item{id}{The node_id to be deleted.}

\item{recursive}{Boolean argument to recursively delete subnodes. Defaults to
FALSE for sake of preventing accidental deletion.}
}
\value{
Boolean TRUE if deletion succeeded.
}
\description{
A function to delete a node on the Open Science Framework. This includes all
types of nodes (e.g., communication, hypothesis, etc.) and includes a full
project. Note that the function recurses throughout subnodes and deletes these as well.
}

