% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{delete.nodes}
\alias{delete.nodes}
\title{Delete a node with its id}
\usage{
delete.nodes(id = NULL, user = NULL, password = NULL)
}
\arguments{
\item{id}{The node_id to be deleted.}

\item{user}{The username to log in with (temporary until OAUTH2.0)}

\item{password}{The password to log in with (temporary until OAUTH2.0)}
}
\value{
Boolean TRUE if deletion succeeded
}
\description{
Delete a node with its id
}

