% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{delete_component}
\alias{delete_component}
\title{Empty out a component and delete it}
\usage{
delete_component(id)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX)}
}
\value{
Boolean, deletion success
}
\description{
This function removes all files contained in a component and then removes
the component itself. NOTE: it does not request confirmation, so please
handle with care. If you do not have backups of the files, it is easy to
lose everything (in the component).
}
\examples{
\dontrun{
delete_component("12345")
}
}
