% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_empty.R
\name{delete_empty}
\alias{delete_empty}
\title{Function to empty out a component and delete it}
\usage{
delete_empty(id, ...)
}
\arguments{
\item{id}{OSF id (osf.io/xxxx)}
}
\value{
Boolean, deletion succeeded?
}
\description{
Function to empty out a component and delete it
}
