% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file based on OSF id}
\usage{
delete_file(id = NULL, ...)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX)}

\item{\ldots}{Additional parameters passed to \code{\link{process_file_id}}}
}
\value{
Boolean, delete succeeded?
}
\description{
Delete a file based on OSF id
}
