% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project from the OSF}
\usage{
delete_project(id, recursive = FALSE, maxdepth = 5)
}
\arguments{
\item{id}{OSF id (osf.io/xxxx)}

\item{recursive}{Boolean, if TRUE will go through folder nesting (see \code{maxdepth})}

\item{maxdepth}{Number of nesting levels to go through}
}
\value{
Boolean, delete succeeded?
}
\description{
Delete a project from the OSF
}
