% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{get.nodes}
\alias{get.nodes}
\title{Retrieve nodes viewable to the authenticated account on the OSF.}
\usage{
get.nodes(id = NULL)
}
\value{
Object dataframe including, for each node:
\enumerate{
\item id
\item title
\item description
\item category
\item date_created
\item date_modified
\item tags
\item registration
\item collection
\item dashboard
\item links
\item public
\item children
\item contributors
\item files
\item node_links
\item parent
\item registrations
}
}
\description{
Retrieve nodes viewable to the authenticated account on the OSF.
}

