% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_nodes.R
\name{get.nodes}
\alias{get.nodes}
\title{Retrieve nodes associated with id}
\usage{
get.nodes(id = NULL, user = NULL, password = NULL, contributors = FALSE,
  files = FALSE, children = FALSE)
}
\arguments{
\item{id}{The id to search for. Use `NULL` to retrieve all,
`me` for logged in account. Maximum of 1 id.}

\item{user}{Username to log in with}

\item{password}{Password to log in with}

\item{contributors}{Boolean to extract the contributors of the node}

\item{files}{Boolean to extract links}
}
\value{
List object of results
}
\description{
Retrieve nodes associated with id
}
\examples{
get.nodes()
get.nodes(id = 'me', user = 'h.schwarzenegger@gmail.com', 'testingtesting')
get.nodes(id = 'nu97z')
}

