% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact_users.R
\name{get.users}
\alias{get.users}
\title{Retrieve all users and their information from the OSF}
\usage{
get.users(id = NULL, user = NULL, password = NULL, nodes = FALSE)
}
\arguments{
\item{id}{The user_id to search for. If NULL (default), returns all users;
if "me" returns logged in user (do not forget to authenticate).}

\item{user}{The username to log in with (temporary until OAUTH2.0)}

\item{password}{The password to log in with (temporary until OAUTH2.0)}
}
\value{
Object dataframe including, for each user:
\enumerate{
\item id
\item fullname
\item given_name
\item middle_names
\item family_name
\item suffix
\item date_registered
\item profile_image_url
\item gitHub
\item scholar
\item personal_website
\item twitter
\item linkedIn
\item impactStory
\item orcid
\item researcherId
\item links
\item nodes
}
}
\description{
Retrieve all users and their information from the OSF
}

