% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config}
\alias{get_config}
\title{Create authorization config}
\usage{
get_config(login_required)
}
\arguments{
\item{login_required}{Boolean}
}
\value{
configuration for use in httr request
}
\description{
Create authorization config
}
