% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{login}
\alias{login}
\title{Login function; interactive without arguments}
\usage{
login(pat = NULL)
}
\arguments{
\item{pat}{Personal Access Token (PAT) for fast login. If no pat is given, function queries for it.}
}
\value{
Personal access token from global environment.
}
\description{
Login function; interactive without arguments
}

