% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{login}
\alias{login}
\title{Login to the OSF}
\usage{
login(pat, store = FALSE)
}
\arguments{
\item{pat}{Personal Access Token (PAT; \url{https://osf.io/settings/tokens})}

\item{store}{If TRUE, store PAT in user's .Renviron}
}
\description{
Login to the OSF API with a Personal Access Token (PAT).
By default and for security reasons, the login lasts as long as the current
session. If you want to store the login across sessions, use argument
\code{store} (this saves the PAT in your .Renviron; see also
\code{\link{logout}}).
}
\examples{
\dontrun{
login(pat = 'this-is-where-you-place-your-pat', store = FALSE)
}
}
