% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{logout}
\alias{logout}
\title{Logout of the OSF}
\usage{
logout()
}
\description{
Remove the Personal Access Token that provides access to the OSF API.
Removes the \code{OSF_PAT} from session environment, not
from the .Renviron file if you stored it with login (that is currently a
manual step).
}
\examples{
\dontrun{
logout()
}
}
