% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-update.R
\name{node-update}
\alias{node-update}
\alias{osf_project_update}
\alias{osf_component_update}
\title{Update an OSF project or component}
\usage{
osf_project_update(id, title = NULL, description = NULL,
  private = NULL)

osf_component_update(id, title = NULL, description = NULL,
  private = NULL)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX)}

\item{title, description}{Change the title or description}

\item{private}{Change privacy settings to private or public}
}
\value{
\code{osf_tbl_node}
}
\description{
Update the title or description of a project or component. You can also
change the privacy settings from public to private (or vice versa).
}
\examples{
\dontrun{
proj <- osf_project_create("temporary-title", private = TRUE)
proj <- osf_project_update(proj$id, title = "Final Title", private = FALSE)
}
}
