% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_download.R
\name{osf_download}
\alias{osf_download}
\title{Download files and directories from OSF}
\usage{
osf_download(x, path = NULL, overwrite = FALSE, decompress = FALSE,
  verbose = FALSE)
}
\arguments{
\item{x}{An \code{\link{osf_tbl_file}} containing a single file or directory.}

\item{path}{Path pointing to a local directory where the downloaded files
will be saved. Default is to use the current working directory.}

\item{overwrite}{Logical, if the local path already exists should it be
replaced with the downloaded file?}

\item{decompress}{Logical, should downloaded directories be automatically
unzipped? Default is \code{TRUE}.}

\item{verbose}{Logical, indicating whether to print informative messages
about interactions with the OSF API (default \code{FALSE}).}
}
\value{
The \code{\link{osf_tbl_file}} input with a new column, \code{"local_path"},
containing the downloaded file's path.
}
\description{
Files stored on OSF can be downloaded locally by providing an
\code{\link{osf_tbl_file}} that contains the OSF file of interest. If the
\code{\link{osf_tbl_file}} contains a directory, a zip file containing the
directory's contents will be downloaded.

By default files are downloaded to your current working directory with the
same filename used on OSF. The \code{path} argument can be used to specify a
different destination \emph{and}, optionally,  a different filename for the
downloaded file. Note, the directory portion of \code{path} must reference an
existing directory.
}
\examples{
\dontrun{
# download a single file
analysis_plan <- osf_retrieve_file("2ryha") \%>\%
  osf_download()

# verify the file was downloaded locally
file.exists(analysis_plan$local_path)
}
}
\seealso{
\itemize{
\item \code{\link[=osf_upload]{osf_upload()}} for uploading files to OSF.
\item \code{\link[=osf_ls_files]{osf_ls_files()}} for listing files and directories on OSF.
}
}
