% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_ls_files.R
\name{osf_ls_files}
\alias{osf_ls_files}
\title{List files and directories on OSF}
\usage{
osf_ls_files(x, path = NULL, type = "any", pattern = NULL,
  n_max = 10)
}
\arguments{
\item{x}{one of the following:
\itemize{
\item an \code{\link{osf_tbl_node}} with a single project or component.
\item an \code{\link{osf_tbl_file}} with a single directory.
}}

\item{path}{list files within the specified subdirectory path}

\item{type}{filter query to include only \code{"file"} or \code{"folder"}}

\item{pattern}{character string used to filter for results that contain the
substring \code{"pattern"} in their name. Note this is a fixed, case-insensitive
search.}

\item{n_max}{Number of results to return from OSF.}
}
\value{
an \code{\link{osf_tbl_file}} with one row for each file or directory
}
\description{
List the files and directories in the top-level of an OSF project, component, or
directory. Specify a \code{path} to list the contents of a particular
subdirectory.
}
\examples{
\dontrun{
# List the files and folders from https://osf.io/gep9v/
project <- osf_retrieve_node("gep9v")
files <- osf_ls_files(project)

# List the files in the first directory
osf_ls_files(files[1, ])
}
}
\seealso{
\code{\link[=osf_ls_nodes]{osf_ls_nodes()}} to generate a list of projects and components
}
