% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{osf_mkdir}
\alias{osf_mkdir}
\title{Create a OSF folder}
\usage{
osf_mkdir(id, path, parent_id = NULL, verbose = FALSE)
}
\arguments{
\item{id}{Parent OSF project id (osf.io/XXXXX; just XXXXX) to create folder in}

\item{path}{Name for the new folder}

\item{parent_id}{Target path for the new folder}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{FALSE})}
}
\value{
Waterbutler URL for folder "root", last subfolder "sub", or all
folders created "all" depanding on the selection input for \code{return}
}
\description{
Creates a new folder in an OSF project or component.
}
\examples{
\dontrun{
proj <- osf_project("Making Directories")
osf_mkdir(proj, "data")
osf_mkdir(proj, "data/raw_data")
osf_mkdir(proj, "super_raw_data", parent_id = "5be5e6a6fe3eca00188178f0")
}
}
