% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_mkdir.R
\name{osf_mkdir}
\alias{osf_mkdir}
\title{Create a OSF folder}
\usage{
osf_mkdir(x, path, verbose = FALSE)
}
\arguments{
\item{x}{an \code{\link{osf_tbl_node}} representing an OSF project or component or an
\code{\link{osf_tbl_file}} containing a directory}

\item{path}{Name for the new directory}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{FALSE})}
}
\value{
an \code{\link{osf_tbl_file}} containing the last directory of the path
}
\description{
Creates a new folder in an OSF project or component.
}
\examples{
\dontrun{
proj <- osf_project_create("Directory Example")
osf_mkdir(proj, "data/raw_data")
}
}
