% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{osf_retrieve}
\alias{osf_retrieve}
\title{Retrieve an entity from OSF based on its identifier}
\usage{
osf_retrieve(id)
}
\arguments{
\item{id}{An OSF identifier corresponding to an OSF user, project, component,
or file.}
}
\value{
an \code{osf_tbl_user}, \code{osf_tbl_node}, or \code{osf_tbl_file}
}
\description{
Use \code{osf_retrieve()} to create an \code{osf_tbl} for an existing OSF entity.
}
\examples{
\dontrun{
 osf_retrieve("dguxh")
}
}
