% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_rm.R
\name{osf_rm}
\alias{osf_rm}
\title{Delete projects or components from OSF}
\usage{
osf_rm(x, recursive = FALSE, verbose = FALSE, check = TRUE)
}
\arguments{
\item{x}{an \code{\link{osf_tbl_node}}}

\item{recursive}{Remove all sub-components before deleting the top-level entity.}

\item{verbose}{Logical, indicating whether to print informative messages
about interactions with the OSF API (default \code{FALSE}).}

\item{check}{If \code{FALSE} deletion will proceed without opening the item or
requesting verification---this effectively removes your safety net.}
}
\value{
Invisibly returns \code{TRUE} if deletion was successful.
}
\description{
Use \code{osf_rm()} to \emph{permanently} delete a project or component from OSF,
including any uploaded files, wiki content, or comments contained therein.
Because this process is irreversible, osfr will first open the item in your
web browser so you can verify the item before proceeding.

If the project or component contains sub-components, those must be deleted
first. Setting \code{recursive = TRUE} will attempt to remove the hierarchy
of sub-components before deleting the top-level entity.

\emph{Note: This functionality is limited to contributors with admin-level
permissions.}
}
\examples{
\dontrun{
project <- osf_create_project("My Short-Lived Project")
osf_rm(project)
}

}
