% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{osf_upload}
\alias{osf_upload}
\title{Upload to OSF}
\usage{
osf_upload(id, path, name = NULL, overwrite = FALSE)
}
\arguments{
\item{id}{OSF project id (osf.io/XXXXX) to upload to.}

\item{path}{Path to file on local machine to upload. Ensure file has
proper extension named (i.e., extension sensitive, not like on Linux)}

\item{name}{Name of the uploaded file (if \code{NULL},
\code{basename(path)} will be used).}
}
\value{
Waterbutler URL
}
\description{
Upload to OSF
}
\seealso{
\code{\link{upload_files}}, \code{\link{upload_revised_files}}
}
