% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{osf_upload}
\alias{osf_upload}
\title{Upload files to OSF}
\usage{
osf_upload(x, path, name = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{Something that identifies a valid OSF upload destination. This can
be one of the following:
\itemize{
\item an \code{osf_tbl_node} with a single project or
component
\item an \code{osf_tbl_file} with a single directory
\item an OSF ID referencing a project, component, or directory
}}

\item{path}{Path to file on local machine to upload. Ensure file has proper
extension named (i.e., extension sensitive, not like on Linux)}

\item{name}{Name of the uploaded file (if \code{NULL}, \code{basename(path)}
will be used).}

\item{overwrite}{Logical, overwrite an existing file with the same name
(default \code{FALSE})? If \code{TRUE}, OSF will automatically update the file and
record the previous version.}
}
\value{
\code{osf_tbl_file}
}
\description{
Upload files to OSF
}
