% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_upload.R
\name{osf_upload}
\alias{osf_upload}
\title{Upload files to OSF}
\usage{
osf_upload(x, path, name = NULL, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{an \code{\link{osf_tbl_node}} with a single project or
component, or an \code{\link{osf_tbl_file}} with a single directory}

\item{path}{Path to file on local machine to upload. Ensure file has proper
extension named (i.e., extension sensitive, not like on Linux)}

\item{name}{Name of the uploaded file (if \code{NULL}, \code{basename(path)}
will be used).}

\item{overwrite}{Logical, overwrite an existing file with the same name
(default \code{FALSE})? If \code{TRUE}, OSF will automatically update the file and
record the previous version.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{FALSE})}
}
\value{
an \code{\link{osf_tbl_file}} containing uploaded file
}
\description{
Upload files to OSF
}
