% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osfr.R
\docType{package}
\name{osfr}
\alias{osfr}
\alias{osfr-package}
\title{osfr: A package for interacting with the Open Science Framework.}
\description{
The osfr package helps you document your research as you go, which
includes creating projects, uploading files, cloning projects to your
desktop in one fell swoop, and much more.
}
\details{
Suggestions for improvements or bug reports always welcome at
https://github.com/chartgerink/osfr
}
