% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{path_file}
\alias{path_file}
\title{Get file download link for the latest version of a file (for direct reading)}
\usage{
path_file(id, private = FALSE)
}
\arguments{
\item{id}{Specify the file id (osf.io/XXXXX)}

\item{private}{Boolean to specify whether file is private}
}
\value{
Return download link (online)
}
\description{
Get file download link for the latest version of a file (for direct reading)
}
\examples{
\dontrun{
read.csv(path_file("5z2bh"))
}
}
