% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file_id.R
\name{process_file_id}
\alias{process_file_id}
\title{Processing a file id to waterbutler}
\usage{
process_file_id(id = NULL, private = FALSE, ...)
}
\arguments{
\item{id}{OSF id (osf.io/xxxx)}

\item{private}{Boolean, if file is private}

\item{\ldots}{Additional parameters passed to \code{\link{construct_link}}}
}
\value{
Waterbutler URL
}
\description{
Processing a file id to waterbutler
}
