% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_type.R
\name{process_type}
\alias{process_type}
\title{Identify type of endpoint for id}
\usage{
process_type(id = NULL, private = FALSE, ...)
}
\arguments{
\item{id}{OSF id to check}

\item{private}{Boolean, in case id is private set to TRUE}

\item{\ldots}{Additional parameters passed to \code{\link{construct_link}}}
}
\value{
Endpoint of id as character (nodes | files)
}
\description{
Identify type of endpoint for id
}
