% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{recurse_node}
\alias{recurse_node}
\title{Function to crawl through OSF project}
\usage{
recurse_node(id = NULL, private = FALSE, maxdepth = 5, path_id = FALSE)
}
\arguments{
\item{id}{OSF parent ID (osf.io/XXXXX) to crawl}

\item{private}{Boolean, search for private too?}

\item{maxdepth}{Integer, amount of levels deep to crawl}

\item{path_id}{Boolean, whether to return paths and ids}
}
\value{
List of OSF ids, with parents as very last.
}
\description{
Function to crawl through OSF project
}
