% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{recurse_node}
\alias{recurse_node}
\title{Function to crawl through OSF project}
\usage{
recurse_node(id, maxdepth = 5)
}
\arguments{
\item{id}{OSF parent ID (osf.io/XXXXX) to crawl}

\item{maxdepth}{Integer, amount of levels deep to crawl}
}
\value{
Character vector of OSF IDs for children of the parent, with the
  parent ID listed last. If there are no child nodes to recurse, the parent
  ID is returned.
}
\description{
Function to crawl through OSF project
}
