% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_nodes}
\alias{search_nodes}
\title{Search OSF nodes}
\usage{
search_nodes(description = NULL, title = NULL, id = NULL, tags = NULL,
  private = FALSE)
}
\arguments{
\item{description}{Search in node description}

\item{title}{Search node titles}

\item{id}{Search for node id}

\item{tags}{Search node tags}

\item{private}{Boolean, search private nodes as well (TRUE) or not (FALSE)}
}
\value{
Data frame of nodes
}
\description{
Search OSF nodes
}
\details{
The categories available are:
\itemize{
\item project
\item hypothesis
\item methods and measures
\item procedure
\item instrumentation
\item data
\item analysis
\item communication
\item other
}
}
