% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{update_component}
\alias{update_component}
\title{Update a component on the OSF}
\usage{
update_component(id, title = NULL, description = NULL,
  private = NULL)
}
\arguments{
\item{id}{OSF id (osf.io/XXXXX; just XXXXX)}

\item{private}{Set project to private/public (default changes to public)}
}
\value{
Boolean, update success
}
\description{
This function updates the private/public status of a component. By default,
the function turns a private component into a public component. This
function simply wraps the \code{\link{update_project}} function
as it has the same operations.
}
\examples{
\dontrun{
update_component("12345")
}
}
\seealso{
\code{\link{update_project}}
}
