% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file to the OSF (both new and revised)}
\usage{
upload_file(id = NULL, filename = NULL, ...)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) to upload to. Specify project to upload new file,
specify a file to upload a revision}

\item{filename}{Local filename to upload}

\item{\ldots}{Additional parameters passed to \code{\link{process_type}}, \code{\link{upload_new}}, and \code{\link{upload_revision}}}
}
\value{
Boolean of upload success
}
\description{
Upload a file to the OSF (both new and revised)
}
\examples{
\dontrun{
upload_file(id = '12345', filename = 'test.pdf')
}
}
