% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_files}
\alias{upload_files}
\title{Upload a file to the OSF (both new and revised)}
\usage{
upload_files(id, path, dest = NULL)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX; just XXXX) to upload to. Specify project to
upload a new file, specify a file to upload a revision.}

\item{path}{Path to file on local machine to upload.}

\item{dest}{Name of the destination file on OSF (if \code{NULL},
\code{basename(path)} will be used). Note that this can be
 used to specify what folder to place files in, e.g. 'my_folder/file.png'.
 Also note that if \code{id} is a file ID, this is not necessary.}
}
\value{
Link if new file created
}
\description{
Upload a file to the OSF (both new and revised)
}
\examples{
\dontrun{
upload_files(id = '12345', path = 'test.pdf')
}
}
