% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_new.R
\name{upload_new}
\alias{upload_new}
\title{Upload a new file to the OSF.}
\usage{
upload_new(id = NULL, filename = NULL, ...)
}
\arguments{
\item{id}{Parent OSF id (osf.io/XXXX) to upload to.}

\item{filename}{Filename on local machine to upload}
}
\value{
Waterbutler URL
}
\description{
Upload a new file to the OSF.
}
\seealso{
\code{\link{upload_file}}, \code{\link{upload_revision}}
}

