% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_revision.R
\name{upload_revision}
\alias{upload_revision}
\title{Upload a revised file to the OSF}
\usage{
upload_revision(id = NULL, filename = NULL, ...)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) of file to revise}

\item{filename}{Local filename to upload as revision}
}
\value{
Boolean of revision success
}
\description{
Upload a revised file to the OSF
}
\seealso{
\code{\link{upload_file}}, \code{\link{upload_new}}
}
