% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{welcome}
\alias{welcome}
\title{Welcome function}
\usage{
welcome(user = NULL, password = NULL)
}
\arguments{
\item{user}{User account}

\item{password}{Password}
}
\value{
Information about OSF version and user in list format
}
\description{
Welcome function
}

