% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{welcome}
\alias{welcome}
\title{Welcome function}
\usage{
welcome(...)
}
\arguments{
\item{...}{other arguments, only for testing.}
}
\value{
Welcome message of logged in user, if any
}
\description{
Welcome function
}

